role attribute added

+++ BaseAndroidProject/app/src/main/java/com/lkb/baseandroidproject/ComModel.kt
@@ -2,8 +2,8 @@ package com.lkb.baseandroidproject
 
 data class ComModel (
    var cmd:String = "",// command to execute corresponding service
-   var cLat:Float=0.0f,// current latitude
-   var cLong:Float=0.0f,// current longitude
+   var cLat: Double =0.0,// current latitude
+   var cLong: Double =0.0,// current longitude
    var cTime:Long=0,// location last update time
    var bats:Int=0// battery status
 )
